split up code, add comments
